Move helper methods and delete filter correctly
